+++ lib/canvas.test.js
@@ -1,7 +1,6 @@
 const test = require('ava')
 const createTestServer = require('create-test-server')
 const proxyquire = require('proxyquire').noPreserveCache()
-const Canvas = require('./canvas')
 
 test.serial('Canvas.getAntagna() returns only users with SIS ID', async t => {
   const server = await createTestServer()
@@ -37,49 +36,3 @@ test.serial('Canvas.getAntagna() returns only the SIS ID', async t => {
 
   t.deepEqual(enrollments, [{ sis_user_id: 'aaa' }])
 })
-
-test.serial(
-  'Canvas.normalizeEnrollments() return objects with all the valid keys',
-  t => {
-    const input = [
-      { user_id: 'user', section_id: 'section', status: 'active', role_id: 12 },
-      {
-        user_id: 'user',
-        xxxx_section_id: 'section',
-        status: 'active',
-        role_id: 12
-      }
-    ]
-
-    const expected = [
-      { user_id: 'user', section_id: 'section', status: 'active', role_id: 12 }
-    ]
-    const result = Canvas.normalizeEnrollments(input)
-
-    t.deepEqual(result, expected)
-  }
-)
-
-test.serial(
-  'Canvas.normalizeEnrollments() return objects without the non-valid keys',
-  t => {
-    const input = [
-      { user_id: 'user', section_id: 'section', status: 'active', role_id: 12 },
-      {
-        user_id: 'user',
-        section_id: 'section2',
-        status: 'active',
-        role_id: 12,
-        role: 'nothing'
-      }
-    ]
-
-    const expected = [
-      { user_id: 'user', section_id: 'section', status: 'active', role_id: 12 },
-      { user_id: 'user', section_id: 'section2', status: 'active', role_id: 12 }
-    ]
-    const result = Canvas.normalizeEnrollments(input)
-
-    t.deepEqual(result, expected)
-  }
-)
