+++ lib/write-file.js
@@ -19,6 +19,7 @@ function normalizeEnrollments (enrollments) {
 }
 
 module.exports = {
+  normalizeEnrollments,
   async writeEnrollments (file, enrollments) {
     const normalized = normalizeEnrollments(enrollments)
